requirements.txt
playwright>=1.40.0
rapidfuzz>=3.5.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
supabase>=2.0.0
python-dotenv>=1.0.0
